refactor: schem'ы вынесены в отдельный пакет
