+++ packages/core/package.json
@@ -13,6 +13,7 @@
     "@ant-design/icons": "^5.3.6",
     "@colibrijs/api-client": "^1.0.0",
     "@colibrijs/module-utils": "^1.0.0",
+    "@colibrijs/schema": "^1.0.0",
     "@colibrijs/types": "^1.0.0",
     "@module-federation/utilities": "^3.0.13",
     "@tanstack/react-query": "^5.32.0",
