+++ packages/schema/index.ts
@@ -1,0 +1,2 @@
+export type { JsonSchema } from './types';
+export { getPropertiesNames } from './utils/get-properties-names';
