+++ packages/types/types.ts
@@ -1,4 +1,3 @@
 export type { IComponent, IComponentConstructorOptions } from './components';
 export type { IElement, IElementConstructorOptions } from './elements';
 export type { IPage, IPageConstructorOptions } from './pages';
-export type { JsonSchema } from './json-schema';
